Added patch for PostgreSQL driver
Fixes ZPS-7424

pg8000 library we use doesn't parse correctly version of posgresql
on Ubuntu system. On Ubuntu starting from PostgreSQL 10+  it returns
version in a following format "Ubuntu 10.2-1.pgdg16.04+1" what library
doesn't handle. As it is not fixed in the latest version of the library
have to manually patch pg8000 lib to handle the case.